Fix conversion warnings.

+++ src/linad99/jacob3.cpp
@@ -173,7 +173,7 @@ void gradient_structure::jacobcalc(int nvar, const uostream& ofs)
     unsigned long int max_last_offset
                = gradient_structure::ARR_LIST1->get_max_last_offset();
 
-    unsigned int size = sizeof(double_and_int );
+    size_t size = sizeof(double_and_int);
 
     for (unsigned int i = 0; i < (max_last_offset/size); i++)
     {
